fix(altera): solve agilex warm reset issue
Agilex warm reset not able to trigger due to the system not able to
detect the magic number. ATF only able to solve for boot core.
For secondary cores, Linux need to update psci driver to WFI the cores
in EL3. Original Linux WFI is EL1. Thus causing secondary cores not working

Signed-off-by: Jit Loon Lim <jit.loon.lim@altera.com>